backEnd/requirements.txt
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comtypes==1.4.11
Flask==3.1.1
flask_cors==3.0.10
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
numpy==2.3.2
pyttsx3==2.99
requests==2.32.4
scikit-learn==1.7.1
scipy==1.16.1
SpeechRecognition==3.14.3
threadpoolctl==3.6.0
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
waitress==2.1.2
psycopg-binary==3.1.18
Flask-Session